refactor: split monolithic index.vue into composable pattern
Extracted 1353-line single file into:
- types.ts — type definitions
- constants.ts — PIECE_CHAR, SVG icons, ICE config
- styles.css — all CSS styles
- composables/useGameEngine.ts — move rules, check detection
- composables/useAI.ts — minimax AI engine
- composables/useBoard.ts — canvas renderer
- composables/useWebRTC.ts — P2P connection management
- index.vue — thin orchestrator (~250 lines)
